% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_qualities.R
\name{decode_qualities}
\alias{decode_qualities}
\title{Decodes ASCII encoded Phred score values}
\usage{
decode_qualities(qualities, offset = 33)
}
\arguments{
\item{qualities}{A single elements  character vector of encoded qualities}

\item{offset}{The phred offset used for}
}
\value{
A numeric vector of Phred scores
}
\description{
Decodes ASCII encoded Phred score values
}
\examples{
decode_qualities("???#;ABAAAHHHHGHFGDHEG@GG@GDGGB>DDDGBDD=")
}
