% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{package}
\name{fastq1R}
\alias{fastq1R}
\title{fastqR: A package to read and manipulate Fastq files}
\description{
This package works with fastq format. It can read this into a tidyverse compatible data structure and
provides functions for manipulating data within them,
}
