% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq.R
\name{read_fastq}
\alias{read_fastq}
\title{Parses a fastq file into tibble}
\usage{
read_fastq(file)
}
\arguments{
\item{file}{A fastq file location}
}
\value{
A tibble of the sections of the entries in the fastq file
}
\description{
Parses a fastq file into tibble
}
\examples{
read_fastq(system.file("good.fq", package="fastq1R"))
}
